Update mutations usage and styles - Use main mutations & delete testing - Use main shims & delete testing - Add basic styles for `<dialog>`s - Make `<header>` full viewport height
