+++ mutations.js
@@ -1,27 +1,63 @@
 import * as handlers from './dataHandlers.js';
 import {$} from './functions.js';
 
-// function lazyLoad(entries, observer) {
-// 	entries.filter(entry => entry.isIntersecting).forEach(async (entry) => {
-// 		try {
-// 			const url = new URL(entry.target.dataset.loadFrom, location.href);
-// 			const resp = await fetch(url);
-// 			if (resp.ok) {
-// 				const parser = new DOMParser();
-// 				const html = await resp.text();
-// 				const content =  parser.parseFromString(html, 'text/html');
-// 				entry.target.append(...content.body.childNodes);
-// 			} else {
-// 				throw new Error(`${resp.url} [${resp.status} ${resp.statusText}]`);
-// 			}
-// 		} catch (error) {
-// 			console.error(error);
-// 		} finally {
-// 			observer.unobserve(entry.target);
-// 			delete entry.target.dataset.loadFrom;
-// 		}
-// 	});
-// }
+function infiniteScroll(entries, observer) {
+	entries.filter(entry => entry.isIntersecting).forEach(async entry => {
+		observer.unobserve(entry.target);
+		const url = new URL(entry.target.dataset.infiniteScroll, location.origin);
+		try {
+			const resp = await fetch(url);
+			if (url.searchParams.has('page')) {
+				url.searchParams.set('page', parseInt(url.searchParams.get('page')) + 1);
+				entry.target.dataset.infiniteScroll = url;
+			}
+			if (resp.ok) {
+				const parser = new DOMParser();
+				const html = await resp.text();
+				const doc = parser.parseFromString(html, 'text/html');
+
+				if (url.hash !== '') {
+					entry.target.before(doc.getElementById(url.hash.substring(1)));
+				} else {
+					entry.target.before(...doc.body.childNodes);
+				}
+				observer.observe(entry.target);
+			} else {
+				throw new Error(`${resp.url} [${resp.status} ${resp.statusText}]`);
+			}
+		} catch (error) {
+			console.error(error);
+			observer.observe(entry.target);
+		}
+	});
+}
+
+function lazyLoad(entries, observer) {
+	entries.filter(entry => entry.isIntersecting).forEach(async entry => {
+		try {
+			const url = new URL(entry.target.dataset.loadFrom, location.href);
+			const resp = await fetch(url);
+			if (resp.ok) {
+				const parser = new DOMParser();
+				const html = await resp.text();
+				const content =  parser.parseFromString(html, 'text/html');
+				entry.target.append(...content.body.childNodes);
+			} else {
+				throw new Error(`${resp.url} [${resp.status} ${resp.statusText}]`);
+			}
+		} catch (error) {
+			console.error(error);
+		} finally {
+			observer.unobserve(entry.target);
+			delete entry.target.dataset.loadFrom;
+		}
+	});
+}
+
+function toggleDetails() {
+	const details = this.closest('details');
+	details.open != details.open;
+}
 
 export const events = {
 	attributes: function() {
@@ -111,7 +147,7 @@ export const options = [
 ];
 
 export function init(base = document.body) {
-	// const observer = new IntersectionObserver(lazyLoad, {rootMargin: '500px 0px 0px 0px'});
+	const observer = new IntersectionObserver(lazyLoad, {rootMargin: '500px 0px 0px 0px'});
 	$('[data-show]', base).click(handlers.show);
 	$('[data-show-modal]', base).click(handlers.showModal);
 	$('[data-close]', base).click(handlers.close);
@@ -121,5 +157,9 @@ export function init(base = document.body) {
 	$('[data-social-share]', base).click(handlers.socialShare);
 	$('[data-fullscreen]', base).click(handlers.fullscreen);
 	$('[data-scroll-to]', base).click(handlers.scrollTo);
-	// $('[data-load-from]', base).each(node => observer.observe(node));
+	$('[data-infinite-scroll]', base).intersect(infiniteScroll);
+	$('[data-load-from]', base).each(node => observer.observe(node));
+	if (document.createElement('details') instanceof HTMLUnknownElement) {
+		$('details > summary').click(toggleDetails);
+	}
 }
