+++ test/index.js
@@ -1,4 +1,4 @@
-import './shims.js';
+import '../shims.js';
 import {loadHandler} from './funcs.js';
 import kbdShortcuts from '../kbd_shortcuts.js';
 import {$} from '../functions.js';
